src/requirements.txt
psycopg2-binary
starlette==0.41.3
databases[postgresql]
fastapi[standard]
uvicorn
httpx
pydantic
sqlmodel
mysqlclient
fuzzywuzzy
dnspython
jellyfish
nicegui
requests
python-Levenshtein
bcrypt
python-multipart
python-dotenv